+++ pkg/store/cache/memcached.go
@@ -15,6 +15,7 @@ import (
 	"github.com/prometheus/prometheus/storage"
 
 	"github.com/thanos-io/thanos/pkg/cacheutil"
+	"github.com/thanos-io/thanos/pkg/tenancy"
 )
 
 const (
@@ -33,18 +34,10 @@ type RemoteIndexCache struct {
 	compressionScheme string
 
 	// Metrics.
-	postingRequests               prometheus.Counter
-	seriesRequests                prometheus.Counter
-	expandedPostingRequests       prometheus.Counter
-	postingHits                   prometheus.Counter
-	seriesHits                    prometheus.Counter
-	expandedPostingHits           prometheus.Counter
-	postingDataSizeBytes          prometheus.Observer
-	expandedPostingDataSizeBytes  prometheus.Observer
-	seriesDataSizeBytes           prometheus.Observer
-	postingsFetchDuration         prometheus.Observer
-	expandedPostingsFetchDuration prometheus.Observer
-	seriesFetchDuration           prometheus.Observer
+	requestTotal  *prometheus.CounterVec
+	hitsTotal     *prometheus.CounterVec
+	dataSizeBytes *prometheus.HistogramVec
+	fetchLatency  *prometheus.HistogramVec
 }
 
 // NewRemoteIndexCache makes a new RemoteIndexCache.
@@ -59,21 +52,23 @@ func NewRemoteIndexCache(logger log.Logger, cacheClient cacheutil.RemoteCacheCli
 		commonMetrics = newCommonMetrics(reg)
 	}
 
-	c.postingRequests = commonMetrics.requestTotal.WithLabelValues(cacheTypePostings)
-	c.seriesRequests = commonMetrics.requestTotal.WithLabelValues(cacheTypeSeries)
-	c.expandedPostingRequests = commonMetrics.requestTotal.WithLabelValues(cacheTypeExpandedPostings)
+	c.requestTotal = commonMetrics.requestTotal
+	c.hitsTotal = commonMetrics.hitsTotal
+	c.dataSizeBytes = commonMetrics.dataSizeBytes
+	c.fetchLatency = commonMetrics.fetchLatency
 
-	c.postingHits = commonMetrics.hitsTotal.WithLabelValues(cacheTypePostings)
-	c.seriesHits = commonMetrics.hitsTotal.WithLabelValues(cacheTypeSeries)
-	c.expandedPostingHits = commonMetrics.hitsTotal.WithLabelValues(cacheTypeExpandedPostings)
+	// Init requestTtotal and hitsTotal with default tenant
+	c.requestTotal.WithLabelValues(cacheTypePostings, tenancy.DefaultTenant)
+	c.requestTotal.WithLabelValues(cacheTypeSeries, tenancy.DefaultTenant)
+	c.requestTotal.WithLabelValues(cacheTypeExpandedPostings, tenancy.DefaultTenant)
 
-	c.postingDataSizeBytes = commonMetrics.dataSizeBytes.WithLabelValues(cacheTypePostings)
-	c.seriesDataSizeBytes = commonMetrics.dataSizeBytes.WithLabelValues(cacheTypeSeries)
-	c.expandedPostingDataSizeBytes = commonMetrics.dataSizeBytes.WithLabelValues(cacheTypeExpandedPostings)
+	c.hitsTotal.WithLabelValues(cacheTypePostings, tenancy.DefaultTenant)
+	c.hitsTotal.WithLabelValues(cacheTypeSeries, tenancy.DefaultTenant)
+	c.hitsTotal.WithLabelValues(cacheTypeExpandedPostings, tenancy.DefaultTenant)
 
-	c.postingsFetchDuration = commonMetrics.fetchLatency.WithLabelValues(cacheTypePostings)
-	c.seriesFetchDuration = commonMetrics.fetchLatency.WithLabelValues(cacheTypeSeries)
-	c.expandedPostingsFetchDuration = commonMetrics.fetchLatency.WithLabelValues(cacheTypeExpandedPostings)
+	c.fetchLatency.WithLabelValues(cacheTypePostings, tenancy.DefaultTenant)
+	c.fetchLatency.WithLabelValues(cacheTypeSeries, tenancy.DefaultTenant)
+	c.fetchLatency.WithLabelValues(cacheTypeExpandedPostings, tenancy.DefaultTenant)
 
 	level.Info(logger).Log("msg", "created index cache")
 
@@ -83,8 +78,8 @@ func NewRemoteIndexCache(logger log.Logger, cacheClient cacheutil.RemoteCacheCli
 // StorePostings sets the postings identified by the ulid and label to the value v.
 // The function enqueues the request and returns immediately: the entry will be
 // asynchronously stored in the cache.
-func (c *RemoteIndexCache) StorePostings(blockID ulid.ULID, l labels.Label, v []byte) {
-	c.postingDataSizeBytes.Observe(float64(len(v)))
+func (c *RemoteIndexCache) StorePostings(blockID ulid.ULID, l labels.Label, v []byte, tenant string) {
+	c.dataSizeBytes.WithLabelValues(cacheTypePostings, tenant).Observe(float64(len(v)))
 	key := cacheKey{blockID.String(), cacheKeyPostings(l), c.compressionScheme}.string()
 	if err := c.memcached.SetAsync(key, v, memcachedDefaultTTL); err != nil {
 		level.Error(c.logger).Log("msg", "failed to cache postings in memcached", "err", err)
@@ -94,8 +89,8 @@ func (c *RemoteIndexCache) StorePostings(blockID ulid.ULID, l labels.Label, v []
 // FetchMultiPostings fetches multiple postings - each identified by a label -
 // and returns a map containing cache hits, along with a list of missing keys.
 // In case of error, it logs and return an empty cache hits map.
-func (c *RemoteIndexCache) FetchMultiPostings(ctx context.Context, blockID ulid.ULID, lbls []labels.Label) (hits map[labels.Label][]byte, misses []labels.Label) {
-	timer := prometheus.NewTimer(c.postingsFetchDuration)
+func (c *RemoteIndexCache) FetchMultiPostings(ctx context.Context, blockID ulid.ULID, lbls []labels.Label, tenant string) (hits map[labels.Label][]byte, misses []labels.Label) {
+	timer := prometheus.NewTimer(c.fetchLatency.WithLabelValues(cacheTypePostings, tenant))
 	defer timer.ObserveDuration()
 
 	keys := make([]string, 0, len(lbls))
@@ -107,7 +102,8 @@ func (c *RemoteIndexCache) FetchMultiPostings(ctx context.Context, blockID ulid.
 	}
 
 	// Fetch the keys from memcached in a single request.
-	c.postingRequests.Add(float64(len(keys)))
+	c.requestTotal.WithLabelValues(cacheTypePostings, tenant).Add(float64(len(keys)))
+
 	results := c.memcached.GetMulti(ctx, keys)
 	if len(results) == 0 {
 		return nil, lbls
@@ -127,16 +123,15 @@ func (c *RemoteIndexCache) FetchMultiPostings(ctx context.Context, blockID ulid.
 
 		hits[lbl] = value
 	}
-
-	c.postingHits.Add(float64(len(hits)))
+	c.hitsTotal.WithLabelValues(cacheTypePostings, tenant).Add(float64(len(hits)))
 	return hits, misses
 }
 
 // StoreExpandedPostings sets the postings identified by the ulid and label to the value v.
 // The function enqueues the request and returns immediately: the entry will be
 // asynchronously stored in the cache.
-func (c *RemoteIndexCache) StoreExpandedPostings(blockID ulid.ULID, keys []*labels.Matcher, v []byte) {
-	c.expandedPostingDataSizeBytes.Observe(float64(len(v)))
+func (c *RemoteIndexCache) StoreExpandedPostings(blockID ulid.ULID, keys []*labels.Matcher, v []byte, tenant string) {
+	c.dataSizeBytes.WithLabelValues(cacheTypeExpandedPostings, tenant).Observe(float64(len(v)))
 	key := cacheKey{blockID.String(), cacheKeyExpandedPostings(labelMatchersToString(keys)), c.compressionScheme}.string()
 
 	if err := c.memcached.SetAsync(key, v, memcachedDefaultTTL); err != nil {
@@ -147,20 +142,20 @@ func (c *RemoteIndexCache) StoreExpandedPostings(blockID ulid.ULID, keys []*labe
 // FetchExpandedPostings fetches multiple postings - each identified by a label -
 // and returns a map containing cache hits, along with a list of missing keys.
 // In case of error, it logs and return an empty cache hits map.
-func (c *RemoteIndexCache) FetchExpandedPostings(ctx context.Context, blockID ulid.ULID, lbls []*labels.Matcher) ([]byte, bool) {
-	timer := prometheus.NewTimer(c.expandedPostingsFetchDuration)
+func (c *RemoteIndexCache) FetchExpandedPostings(ctx context.Context, blockID ulid.ULID, lbls []*labels.Matcher, tenant string) ([]byte, bool) {
+	timer := prometheus.NewTimer(c.fetchLatency.WithLabelValues(cacheTypeExpandedPostings, tenant))
 	defer timer.ObserveDuration()
 
 	key := cacheKey{blockID.String(), cacheKeyExpandedPostings(labelMatchersToString(lbls)), c.compressionScheme}.string()
 
 	// Fetch the keys from memcached in a single request.
-	c.expandedPostingRequests.Add(1)
+	c.requestTotal.WithLabelValues(cacheTypeExpandedPostings, tenant).Add(1)
 	results := c.memcached.GetMulti(ctx, []string{key})
 	if len(results) == 0 {
 		return nil, false
 	}
 	if res, ok := results[key]; ok {
-		c.expandedPostingHits.Add(1)
+		c.hitsTotal.WithLabelValues(cacheTypeExpandedPostings, tenant).Add(1)
 		return res, true
 	}
 	return nil, false
@@ -169,8 +164,8 @@ func (c *RemoteIndexCache) FetchExpandedPostings(ctx context.Context, blockID ul
 // StoreSeries sets the series identified by the ulid and id to the value v.
 // The function enqueues the request and returns immediately: the entry will be
 // asynchronously stored in the cache.
-func (c *RemoteIndexCache) StoreSeries(blockID ulid.ULID, id storage.SeriesRef, v []byte) {
-	c.seriesDataSizeBytes.Observe(float64(len(v)))
+func (c *RemoteIndexCache) StoreSeries(blockID ulid.ULID, id storage.SeriesRef, v []byte, tenant string) {
+	c.dataSizeBytes.WithLabelValues(cacheTypeSeries, tenant).Observe(float64(len(v)))
 	key := cacheKey{blockID.String(), cacheKeySeries(id), ""}.string()
 
 	if err := c.memcached.SetAsync(key, v, memcachedDefaultTTL); err != nil {
@@ -181,8 +176,8 @@ func (c *RemoteIndexCache) StoreSeries(blockID ulid.ULID, id storage.SeriesRef, 
 // FetchMultiSeries fetches multiple series - each identified by ID - from the cache
 // and returns a map containing cache hits, along with a list of missing IDs.
 // In case of error, it logs and return an empty cache hits map.
-func (c *RemoteIndexCache) FetchMultiSeries(ctx context.Context, blockID ulid.ULID, ids []storage.SeriesRef) (hits map[storage.SeriesRef][]byte, misses []storage.SeriesRef) {
-	timer := prometheus.NewTimer(c.seriesFetchDuration)
+func (c *RemoteIndexCache) FetchMultiSeries(ctx context.Context, blockID ulid.ULID, ids []storage.SeriesRef, tenant string) (hits map[storage.SeriesRef][]byte, misses []storage.SeriesRef) {
+	timer := prometheus.NewTimer(c.fetchLatency.WithLabelValues(cacheTypeSeries, tenant))
 	defer timer.ObserveDuration()
 
 	keys := make([]string, 0, len(ids))
@@ -194,7 +189,7 @@ func (c *RemoteIndexCache) FetchMultiSeries(ctx context.Context, blockID ulid.UL
 	}
 
 	// Fetch the keys from memcached in a single request.
-	c.seriesRequests.Add(float64(len(ids)))
+	c.requestTotal.WithLabelValues(cacheTypeSeries, tenant).Add(float64(len(ids)))
 	results := c.memcached.GetMulti(ctx, keys)
 	if len(results) == 0 {
 		return nil, ids
@@ -214,8 +209,7 @@ func (c *RemoteIndexCache) FetchMultiSeries(ctx context.Context, blockID ulid.UL
 
 		hits[id] = value
 	}
-
-	c.seriesHits.Add(float64(len(hits)))
+	c.hitsTotal.WithLabelValues(cacheTypeSeries, tenant).Add(float64(len(hits)))
 	return hits, misses
 }